+++ models/ProductTag.js
@@ -7,20 +7,20 @@ class ProductTag extends Model {}
 ProductTag.init(
   {
     id: {
-      type: DataTypes.INTEGER(100),
+      type: DataTypes.INTEGER(32),
       allowNull: false,
       primaryKey: true,
       autoIncrement: true,
     },
     product_id: {
-      type: DataTypes.INTEGER(100),
+      type: DataTypes.INTEGER(32),
       references: {
         model: "product",
         key: "id",
       },
     },
     tag_id: {
-      type: DataTypes.INTEGER(100),
+      type: DataTypes.INTEGER(32),
       references: {
         model: "tag",
         key: "id",
